SPlit application into infrastructure and domain layer.
