+++ src/DomainModel/Repository/MarsResearchStationRepository.php
@@ -1,0 +1,8 @@
+<?php
+declare(strict_types=1);
+
+namespace App\DomainModel\Repository;
+
+interface MarsResearchStationRepository
+{
+}
