--- src/Repository/DeliveryRepository.php
@@ -1,22 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace App\Repository;
-
-use App\Entity\Delivery;
-use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
-use Doctrine\Persistence\ManagerRegistry;
-
-/**
- * @method Delivery|null find($id, $lockMode = null, $lockVersion = null)
- * @method Delivery|null findOneBy(array $criteria, array $orderBy = null)
- * @method Delivery[]    findAll()
- * @method Delivery[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
- */
-class DeliveryRepository extends ServiceEntityRepository
-{
-    public function __construct(ManagerRegistry $registry)
-    {
-        parent::__construct($registry, Delivery::class);
-    }
-}
