--- src/Repository/ExpeditionRepository.php
@@ -1,21 +1,0 @@
-<?php
-
-namespace App\Repository;
-
-use App\Entity\Expedition;
-use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
-use Doctrine\Persistence\ManagerRegistry;
-
-/**
- * @method Expedition|null find($id, $lockMode = null, $lockVersion = null)
- * @method Expedition|null findOneBy(array $criteria, array $orderBy = null)
- * @method Expedition[]    findAll()
- * @method Expedition[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
- */
-class ExpeditionRepository extends ServiceEntityRepository
-{
-    public function __construct(ManagerRegistry $registry)
-    {
-        parent::__construct($registry, Expedition::class);
-    }
-}
