--- src/Repository/MarsResearchStationRepository.php
@@ -1,22 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace App\Repository;
-
-use App\Entity\MarsResearchStation;
-use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
-use Doctrine\Persistence\ManagerRegistry;
-
-/**
- * @method MarsResearchStation|null find($id, $lockMode = null, $lockVersion = null)
- * @method MarsResearchStation|null findOneBy(array $criteria, array $orderBy = null)
- * @method MarsResearchStation[]    findAll()
- * @method MarsResearchStation[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
- */
-class MarsResearchStationRepository extends ServiceEntityRepository
-{
-    public function __construct(ManagerRegistry $registry)
-    {
-        parent::__construct($registry, MarsResearchStation::class);
-    }
-}
