--- src/Repository/SpaceResearchStationRepository.php
@@ -1,22 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace App\Repository;
-
-use App\Entity\SpaceResearchStation;
-use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
-use Doctrine\Persistence\ManagerRegistry;
-
-/**
- * @method SpaceResearchStation|null find($id, $lockMode = null, $lockVersion = null)
- * @method SpaceResearchStation|null findOneBy(array $criteria, array $orderBy = null)
- * @method SpaceResearchStation[]    findAll()
- * @method SpaceResearchStation[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
- */
-class SpaceResearchStationRepository extends ServiceEntityRepository
-{
-    public function __construct(ManagerRegistry $registry)
-    {
-        parent::__construct($registry, SpaceResearchStation::class);
-    }
-}
